Trying FullyFusedMLP with FP32

+++ nsff_exp/models/fused_nerf.py
@@ -87,9 +87,6 @@ class FusedDynamicNeRF(nn.Module):
 
     @nvtx.annotate("Fused Dynamic NeRF forward")
     def forward(self, x):
-        # Fused NeRF is 16-bit internally
-        x = x.to(torch.float16)
-
         # 4 input channels, 3 view channels
         input_position, input_view = x.split([4, 3], dim=-1)
         x = self.density_mlp(input_position)
@@ -97,12 +94,13 @@ class FusedDynamicNeRF(nn.Module):
         # 2 x 3-dim scene flow, 2 x 1-dim disocclusion blend, 128-dim feature vector
         scene_flow, disocclusion_blend, feature_vector = torch.split(x, [6, 2, self.W], dim=-1)
 
-        scene_flow = torch.tanh(scene_flow.to(torch.float32))
-        disocclusion_blend = torch.sigmoid(disocclusion_blend.to(torch.float32))
-        density = feature_vector[:, 0:1].to(torch.float32)
+        scene_flow = torch.tanh(scene_flow)
+        disocclusion_blend = torch.sigmoid(disocclusion_blend)
+        density = feature_vector[:, 0:1]
 
         rgb = self.color_mlp(torch.cat([input_view, feature_vector], dim=-1))
-        rgb = rgb.to(torch.float32)
+
+        import pdb; pdb.set_trace()
 
         return torch.cat([rgb, density, scene_flow, disocclusion_blend], dim=-1)
 
@@ -118,9 +116,6 @@ class FusedStaticNeRF(nn.Module):
 
     @nvtx.annotate("Fused Static NeRF forward")
     def forward(self, x):
-        # Fused NeRF is 16-bit internally
-        x = x.to(torch.float16)
-
         # 3 input channels, 3 view channels
         input_position, input_view = x.split([3, 3], dim=-1)
         x = self.density_mlp(input_position)
@@ -128,10 +123,9 @@ class FusedStaticNeRF(nn.Module):
         # 1-dim blending weight, 128-dim feature vector
         blending, feature_vector = x.split([1, self.W], dim=-1)
 
-        blending = torch.sigmoid(blending.to(torch.float32))
-        density = feature_vector[:, 0:1].to(torch.float32)
+        blending = torch.sigmoid(blending)
+        density = feature_vector[:, 0:1]
 
         rgb = self.color_mlp(torch.cat([input_view, feature_vector], dim=-1))
-        rgb = rgb.to(torch.float32)
 
         return torch.cat([rgb, density, blending], dim=-1)
